Pin sprockets gem to < v4.0
Updating Rails to v6.1.0 causes sprockets to be updated to v4.0.2 as a
dependency. That version expects app/assets/config/manifest.js to be
present, which it is not since switching to webpacker. This causes an
error and prevents the server starting.

This commit adds sprockets to the gemfile and pins it as < v4.0 to
prevent any unexpected updates in future.

See https://github.com/rails/sprockets-rails/issues/444 for discussion
of the issue.